+++ src/main/java/com/mageddo/dnsproxyserver/dagger/Factory.java
@@ -1,7 +1,7 @@
 package com.mageddo.dnsproxyserver.dagger;
 
-import com.mageddo.dnsproxyserver.dns.server.ServerStarter;
-import com.mageddo.dnsproxyserver.dns.server.solver.Solver;
+import com.mageddo.dnsproxyserver.server.dns.ServerStarter;
+import com.mageddo.dnsproxyserver.server.dns.solver.Solver;
 import dagger.Component;
 
 import javax.inject.Singleton;
