+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/solver/DockerSolver.java
@@ -1,4 +1,4 @@
-package com.mageddo.dnsproxyserver.dns.server.solver;
+package com.mageddo.dnsproxyserver.server.dns.solver;
 
 import com.mageddo.dnsproxyserver.docker.DockerRepository;
 import com.mageddo.dnsproxyserver.utils.Ips;
