+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/solver/RemoteSolver.java
@@ -1,4 +1,4 @@
-package com.mageddo.dnsproxyserver.dns.server.solver;
+package com.mageddo.dnsproxyserver.server.dns.solver;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
